Feat stopplan (#223)
* chore: stop plan on delete

+++ arex-web-api/src/main/java/com/arextest/web/api/service/controller/ReportQueryController.java
@@ -4,7 +4,10 @@ import javax.annotation.Resource;
 import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 
+import com.arextest.web.common.HttpUtils;
 import org.apache.commons.lang3.StringUtils;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -109,6 +112,8 @@ public class ReportQueryController {
     private SceneReportService sceneReportService;
     @Resource
     private RecordService recordService;
+    @Value("${arex.schedule.stop.url}")
+    private String stopPlanUrl;
 
     @Deprecated
     @PostMapping("/pushCompareResults")
@@ -270,9 +275,20 @@ public class ReportQueryController {
     @GetMapping("/delete/{planId}")
     @ResponseBody
     public Response deleteReport(@PathVariable String planId) {
-        SuccessResponseType response = new SuccessResponseType();
-        response.setSuccess(reportService.deleteReport(planId));
-        return ResponseUtils.successResponse(response);
+        try {
+            SuccessResponseType response = new SuccessResponseType();
+            ResponseEntity<String> stopRes = HttpUtils.get(stopPlanUrl + "/?planId=" + planId, String.class);
+            if (!stopRes.hasBody() || StringUtils.isEmpty(stopRes.getBody())) {
+                LOGGER.error("stop plan error, planId:{}", planId);
+                return ResponseUtils.errorResponse("stop plan error", ResponseCode.REQUESTED_HANDLE_EXCEPTION);
+            }
+            LOGGER.info("stop plan success, {}", stopRes.getBody());
+            response.setSuccess(reportService.deleteReport(planId));
+            return ResponseUtils.successResponse(response);
+        } catch (Throwable t) {
+            LOGGER.error("delete plan error", t);
+            return ResponseUtils.errorResponse(t.getMessage(), ResponseCode.REQUESTED_HANDLE_EXCEPTION);
+        }
     }
 
     @GetMapping("/querySceneInfo/{planId}/{planItemId}")
